refactoring search + new search helper
