+++ projects/bp-gallery/src/views/gallery/search/SearchView.tsx
@@ -7,7 +7,8 @@ import './SearchView.scss';
 import SearchHub from './searchHub/SearchHub';
 import PictureScrollGrid from '../shared/PictureScrollGrid';
 import SearchBreadcrumbs from './SearchBreadcrumbs';
-import { PictureFiltersInput } from '../../../graphql/APIConnector';
+//import { PictureFiltersInput } from '../../../graphql/APIConnector';
+import { convertSearchParamsToPictureFilters } from '../helpers/search-filters';
 import {
   Button,
   IconButton,
@@ -80,203 +81,6 @@ export const addNewParamToSearchPath = (
   });
   return { searchVal: asSearchPath(searchParams), isValid: true };
 };
-
-const paramToTimefilter = (timeParam: string) => {
-  console.log(timeParam);
-  // if (searchParams.get(SearchType.DECADE) === 'pre50') searchParams.set(SearchType.DECADE, '40');
-  const year = parseInt(timeParam);
-
-  if (!isNaN(year)) {
-    let startTime = `${year}-01-01T00:00:00Z`;
-    let endTime = `${year}-12-31T23:59:59Z`;
-    if (year < 100) {
-      startTime = year === 40 ? '1900-01-01T00:00:00Z' : `19${year}-01-01T00:00:00Z`;
-      endTime = `19${year}-12-31T23:59:59Z`;
-    }
-
-    const time_range_tag_filter = {
-      start: {
-        gte: startTime,
-      },
-      end: {
-        lte: endTime,
-      },
-    };
-
-    return {
-      or: [
-        {
-          time_range_tag: time_range_tag_filter,
-        },
-        {
-          verified_time_range_tag: time_range_tag_filter,
-        },
-      ],
-    };
-  }
-};
-
-export const convertSearchParamsToPictureFilters = (searchParams: URLSearchParams) => {
-  const filters: PictureFiltersInput = { and: [], or: [] };
-
-  if (searchParams.has(SearchType.ALL)) {
-    const params = searchParams.getAll(SearchType.ALL).map(decodeURIComponent);
-    params.forEach(function (value) {
-      if (value === 'pre50') {
-        value = '40';
-      }
-
-      const keyword = decodeURIComponent(value);
-      const keyword_tag_filter = {
-        name: {
-          containsi: keyword,
-        },
-      };
-
-      const q = decodeURIComponent(value);
-
-      const value_number = parseInt(value);
-      let time_range_tag_filter = {};
-      if (!isNaN(value_number)) {
-        if (value_number.toString().startsWith('19') && value_number.toString().length === 4) {
-          const startTime = `19${parseInt(value_number.toString().substring(2))}-01-01T00:00:00Z`;
-          const endTime = `19${parseInt(value_number.toString().substring(2))}-12-31T23:59:59Z`;
-
-          time_range_tag_filter = {
-            start: {
-              gte: startTime,
-            },
-            end: {
-              lte: endTime,
-            },
-          };
-        } else if (value_number.toString().length === 2) {
-          const startTime =
-            value_number === 40 ? '1900-01-01T00:00:00Z' : `19${value_number}-01-01T00:00:00Z`;
-          const endTime = `19${value_number}-12-31T23:59:59Z`;
-
-          time_range_tag_filter = {
-            start: {
-              gte: startTime,
-            },
-            end: {
-              lte: endTime,
-            },
-          };
-        }
-      }
-
-      filters.and?.push({
-        or: [
-          {
-            keyword_tags: keyword_tag_filter,
-          },
-          {
-            verified_keyword_tags: keyword_tag_filter,
-          },
-          {
-            descriptions: {
-              text: {
-                containsi: q,
-              },
-            },
-          },
-          {
-            time_range_tag: time_range_tag_filter,
-          },
-          {
-            verified_time_range_tag: time_range_tag_filter,
-          },
-        ],
-      });
-    });
-  }
-  if (searchParams.has(SearchType.DECADE)) {
-    const timeParams = searchParams.getAll(SearchType.DECADE);
-    timeParams.forEach(timeParam => {
-      filters.and?.push(paramToTimefilter(timeParam) as PictureFiltersInput);
-    });
-  }
-
-  if (searchParams.has(SearchType.KEYWORD)) {
-    const keywords = searchParams.getAll(SearchType.KEYWORD).map(decodeURIComponent);
-    keywords.forEach((keyword: string) => {
-      const keyword_tag_filter = {
-        name: {
-          containsi: keyword,
-        },
-      };
-      filters.and?.push({
-        or: [
-          {
-            keyword_tags: keyword_tag_filter,
-          },
-          {
-            verified_keyword_tags: keyword_tag_filter,
-          },
-        ],
-      });
-    });
-  }
-
-  if (searchParams.has(SearchType.DESCRIPTION)) {
-    const q = searchParams.getAll(SearchType.DESCRIPTION).map(decodeURIComponent);
-    q.forEach((param: string) => {
-      filters.and?.push({
-        descriptions: {
-          text: {
-            containsi: param,
-          },
-        },
-      });
-    });
-  }
-
-  if (searchParams.has(SearchType.PERSON)) {
-    const persons = searchParams.getAll(SearchType.PERSON).map(decodeURIComponent);
-    persons.forEach((person: string) => {
-      const person_tag_filter = {
-        name: {
-          containsi: person,
-        },
-      };
-      filters.and?.push({
-        or: [
-          {
-            person_tags: person_tag_filter,
-          },
-          {
-            verified_person_tags: person_tag_filter,
-          },
-        ],
-      });
-    });
-  }
-
-  if (searchParams.has(SearchType.LOCATION)) {
-    const locations = searchParams.getAll(SearchType.LOCATION).map(decodeURIComponent);
-    locations.forEach((location: string) => {
-      const location_tag_filter = {
-        name: {
-          containsi: location,
-        },
-      };
-      filters.and?.push({
-        or: [
-          {
-            location_tags: location_tag_filter,
-          },
-          {
-            verified_location_tags: location_tag_filter,
-          },
-        ],
-      });
-    });
-  }
-
-  return filters;
-};
-
 
 const SearchView = ({ scrollPos, scrollHeight }: { scrollPos: number; scrollHeight: number }) => {
   const [searchSnippet, setSearchSnippet] = useState<string>('');
